requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
dataclasses==0.6
dj-database-url==2.1.0
Django==4.1.11
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
idna==2.7
marvel==0.1.0
PyJWT==2.8.0
pymongo==3.12.3
python-decouple==3.8
pytz==2023.3.post1
requests==2.20.0
rest-framework-simplejwt==0.0.2
sqlparse==0.2.4
typing_extensions==4.7.1
urllib3==1.24.3
